+++ tests/end2end/test_zhmc_partition_list.py
@@ -63,54 +63,72 @@ def assert_partition_list(partition_list, exp_partition_dict):
     Assert the output of the zhmc_partition_list module
 
     Parameters:
-      partition_list(list): Result of zhmc_partition_list module, a list
-        of partition properties as documented (in HMC notation with dashes).
-      exp_partition_dict(dict): Expected properties for each expected result
-        item. Key: tuple(CPC name, partition name), Value: All properties
-        of the partition plus artificial properties (in HMC notation with
-        dashes).
+
+      partition_list(list): Result of zhmc_partition_list module, as a list of
+        dicts of partition properties as documented (with underscores in their
+        names).
+
+      exp_partition_dict(dict): Expected partitions with their properties.
+        Key: tuple(CPC name, partition name).
+        Value: Dict of expected partition properties (including any artificial
+        properties, all with underscores in their names).
     """
 
     assert isinstance(partition_list, list)
 
-    exp_cpc_part_names = list(exp_partition_dict)
-    cpc_part_names = [(pi.get('cpc_name', None), pi.get('name', None))
-                      for pi in partition_list]
-    assert set(cpc_part_names) == set(exp_cpc_part_names)
+    exp_part_keys = list(exp_partition_dict.keys())
+    part_keys = [(ri.get('cpc_name', None), ri.get('name', None))
+                 for ri in partition_list]
+    assert set(part_keys) == set(exp_part_keys)
 
     for partition_item in partition_list:
         partition_name = partition_item.get('name', None)
         cpc_name = partition_item.get('cpc_name', None)
-        cpc_part_name = (cpc_name, partition_name)
+        part_key = (cpc_name, partition_name)
 
         assert partition_name is not None, \
             "Returned partition {pi!r} does not have a 'name' property". \
             format(pi=partition_item)
 
-        assert cpc_part_name in exp_partition_dict, \
-            "Result contains unexpected partition {p!r} in CPC {c!r}". \
-            format(p=partition_name, c=cpc_name)
+        assert cpc_name is not None, \
+            "Returned partition {pi!r} does not have a 'cpc_name' property". \
+            format(pi=partition_item)
 
-        exp_partition_props = exp_partition_dict[cpc_part_name]
+        assert part_key in exp_partition_dict, \
+            "Result contains unexpected partition {pn!r} in CPC {cn!r}". \
+            format(pn=partition_name, cn=cpc_name)
 
+        exp_partition_properties = exp_partition_dict[part_key]
         for pname, pvalue in partition_item.items():
-
-            # Verify normal properties
-            pname_hmc = pname.replace('_', '-')
-            assert pname_hmc in exp_partition_props, \
-                "Unexpected property {pn!r} in result partition {rn!r}". \
-                format(pn=pname_hmc, rn=partition_name)
-            exp_value = exp_partition_props[pname_hmc]
+            assert '-' not in pname, \
+                "Property {pn!r} in partition {rn!r} is returned with " \
+                "hyphens in the property name". \
+                format(pn=pname, rn=partition_name)
+            assert pname in exp_partition_properties, \
+                "Unexpected property {pn!r} in result partition {rn!r}. " \
+                "Expected properties: {en!r}". \
+                format(pn=pname, rn=partition_name,
+                       en=list(exp_partition_properties.keys()))
+            exp_value = exp_partition_properties[pname]
             assert pvalue == exp_value, \
                 "Incorrect value for property {pn!r} of result partition " \
                 "{rn!r}". \
-                format(pn=pname_hmc, rn=partition_name)
+                format(pn=pname, rn=partition_name)
 
 
 @pytest.mark.parametrize(
     "with_cpc", [
         pytest.param(False, id="with_cpc=False"),
         pytest.param(True, id="with_cpc=True"),
+    ]
+)
+@pytest.mark.parametrize(
+    "property_flags", [
+        pytest.param({}, id="property_flags()"),
+        pytest.param({'additional_properties': ['description']},
+                     id="property_flags(additional_properties=[description])"),
+        pytest.param({'full_properties': True},
+                     id="property_flags(full_properties=True)"),
     ]
 )
 @pytest.mark.parametrize(
@@ -121,7 +139,8 @@ def assert_partition_list(partition_list, exp_partition_dict):
 )
 @mock.patch("plugins.modules.zhmc_partition_list.AnsibleModule", autospec=True)
 def test_zhmc_partition_list(
-        ansible_mod_cls, check_mode, with_cpc, dpm_mode_cpcs):  # noqa: F811, E501
+        ansible_mod_cls, check_mode, property_flags, with_cpc,
+        dpm_mode_cpcs):  # noqa: F811, E501
     """
     Test the zhmc_partition_list module with DPM mode CPCs.
     """
@@ -142,20 +161,27 @@ def test_zhmc_partition_list(
 
         faked_session = session if hd.mock_file else None
 
+        full_properties = property_flags.get('full_properties', False)
+        additional_properties = property_flags.get(
+            'additional_properties', [])
+
         # Determine the expected partitions on the HMC
-        if DEBUG:
-            print("Debug: Listing expected partitions")
         hmc_version = client.query_api_version()['hmc-version']
         hmc_version_info = [int(x) for x in hmc_version.split('.')]
-        if hmc_version_info < [2, 14, 0]:
+        if hmc_version_info < [2, 14, 0] or additional_properties:
             # List the LPARs in the traditional way
             if with_cpc:
-                exp_partitions = cpc.partitions.list()
+                exp_partitions = cpc.partitions.list(
+                    additional_properties=additional_properties,
+                    full_properties=full_properties)
             else:
                 cpcs_ = client.cpcs.list()
                 exp_partitions = []
                 for cpc_ in cpcs_:
-                    exp_partitions.extend(cpc_.partitions.list())
+                    _partitions = cpc_.partitions.list(
+                        additional_properties=additional_properties,
+                        full_properties=full_properties)
+                    exp_partitions.extend(_partitions)
         else:
             # List the LPARs using the new operation
             if with_cpc:
@@ -163,14 +189,12 @@ def test_zhmc_partition_list(
             else:
                 filter_args = None
             exp_partitions = console.list_permitted_partitions(
-                filter_args=filter_args)
+                filter_args=filter_args,
+                full_properties=full_properties)
+
         exp_partition_dict = {}
         se_versions = {}
         for partition in exp_partitions:
-            if DEBUG:
-                print("Debug: Getting expected properties of partition {p!r} "
-                      "on CPC {c!r}".format(p=partition.name, c=cpc.name))
-            partition.pull_full_properties()
             cpc = partition.manager.parent
             try:
                 se_version = se_versions[cpc.name]
@@ -180,18 +204,28 @@ def test_zhmc_partition_list(
                           format(c=cpc.name))
                 se_version = cpc.get_property('se-version')
                 se_versions[cpc.name] = se_version
-            exp_properties = {}
-            exp_properties.update(partition.properties)
-            exp_properties['cpc-name'] = cpc.name
-            exp_properties['se-version'] = se_version
-            exp_cpc_part_name = (cpc.name, partition.name)
-            exp_partition_dict[exp_cpc_part_name] = exp_properties
+            exp_properties = {
+                'cpc_name': cpc.name,
+                'se_version': se_version,
+            }
+            for pname_hmc, pvalue in partition.properties.items():
+                pname = pname_hmc.replace('-', '_')
+                exp_properties[pname] = pvalue
+            if DEBUG:
+                print("Debug: Expected properties of partition {p!r} "
+                      "on CPC {c!r}: {n!r}".
+                      format(p=partition.name, c=cpc.name,
+                             n=list(exp_properties.keys())))
+            exp_part_key = (cpc.name, partition.name)
+            exp_partition_dict[exp_part_key] = exp_properties
 
         # Prepare module input parameters (must be all required + optional)
         params = {
             'hmc_host': hmc_host,
             'hmc_auth': hmc_auth,
             'cpc_name': cpc.name if with_cpc else None,
+            'additional_properties': additional_properties,
+            'full_properties': full_properties,
             'log_file': LOG_FILE,
             '_faked_session': faked_session,
         }